feat: add TablesModule — GET /tables/ranking
Replaces 6 legacy table endpoints (top-sales-reps, branch-ranking,
product-details, product-detailed-performance,
branch-detailed-performance, top-customers) with one
?entity=cashier|branch|product|customer route. Each entity keeps its
own clearly-written query (they genuinely have different useful
columns) rather than forcing a fake generic template — the win is one
route/DTO/RBAC/cache config, not one query shape.

Also fixes a real bug this surfaced: COUNT(DISTINCT ...) returns
Postgres BIGINT, which deserializes as a JS BigInt that neither
JSON.stringify nor the Redis cache serializer can handle — 500s on
any row that actually has data, silently fine on an empty table
(zero matching rows, not a zero to serialize). Cast every COUNT to
::int here and in ChartsService's orders metric, which had the same
latent bug untested until now. Added an e2e test seeding real rows
specifically to catch this class of thing going forward — an empty
database was hiding it.

Not yet covered (follow-up, same as Charts): customer-segmentation-table
dropped as redundant with the future KDS Analytics customer-segmentation
endpoint; price/cost YoY comparison and region-cost tables (4 remaining
legacy endpoints) still to come.

+++ apps/api/src/charts/charts.service.ts
@@ -14,7 +14,11 @@ const TREND_METRIC_EXPR: Record<TrendMetric, Prisma.Sql> = {
   [TrendMetric.COST]: Prisma.sql`COALESCE(SUM(ri.total_cost), 0)`,
   [TrendMetric.PROFIT]: Prisma.sql`COALESCE(SUM(ri.total_margin), 0)`,
   [TrendMetric.QUANTITY]: Prisma.sql`COALESCE(SUM(ri.quantity), 0)`,
-  [TrendMetric.ORDERS]: Prisma.sql`COUNT(DISTINCT r.id)`,
+  // Postgres COUNT() returns BIGINT, which node-pg/Prisma surface as a JS
+  // BigInt — JSON.stringify (and therefore both the HTTP response and the
+  // Redis cache serializer) cannot serialize that. Casting to ::int avoids
+  // it; receipt counts will never approach the int4 ceiling.
+  [TrendMetric.ORDERS]: Prisma.sql`COUNT(DISTINCT r.id)::int`,
 };
 
 const TREND_GRANULARITY_EXPR: Record<TrendGranularity, Prisma.Sql> = {
